Terms: Update description generator for new Faker behavior
With the new version of the Faker library introduced in #147, the
behavior of the `realTextBetween` seems to have changed, making it more
strict in its interpretation of the input parameters. This change avoids
fatal errors by ensuring that there is always a range of 40 between the
specified minimum number of characters and the maximum, so that the
generator can always build a string with a length that falls somewhere
within the range.